Add support for puppetlabs-apt 2.0+
Note: This is a temporary stopgap solution! We are doing this to enable
upgrading to puppetlabs-apt 2.0+ in opentable/puppet-modules without
requiring a full upgrade of puppetlabs-mongodb. Our expectation is that
Chuck D. will work in parallel to upgrade us to 0.13.0 or better of this
module which already includes the changes in this commit.

If this fork remains in production beyond Apr 30, we have failed as the
maintainers of Puppet. :)